Remove the dynamodb code
