+++ src/emma_experience_hub/api/clients/__init__.py
@@ -3,7 +3,6 @@ from emma_experience_hub.api.clients.compound_splitter import CompoundSplitterCl
 from emma_experience_hub.api.clients.confirmation_response_classifier import (
     ConfirmationResponseClassifierClient,
 )
-from emma_experience_hub.api.clients.dynamo_db import DynamoDbClient
 from emma_experience_hub.api.clients.emma_policy import EmmaPolicyClient
 from emma_experience_hub.api.clients.feature_extractor import FeatureExtractorClient
 from emma_experience_hub.api.clients.out_of_domain_detector import OutOfDomainDetectorClient
