+++ src/emma_experience_hub/api/clients/simbot/session_db.py
@@ -1,88 +1,120 @@
+import sqlite3
 from concurrent.futures import ThreadPoolExecutor
-from typing import Any
+from pathlib import Path
 
-import boto3
-from boto3.dynamodb.conditions import Key
-from botocore.exceptions import ClientError
 from loguru import logger
 
-from emma_experience_hub.api.clients.dynamo_db import DynamoDbClient
 from emma_experience_hub.datamodels.simbot import SimBotSessionTurn
 
 
-class SimBotSessionDbClient(DynamoDbClient):
-    """Client for storing SimBot session data."""
+class SimBotSessionDbClient:
+    """Local Client for storing SimBot session data."""
 
-    primary_key = "session_id"
-    sort_key = "idx"
-    data_key = "turn"
+    primary_key: str
+    sort_key: str
+    data_key: str
+
+    def __init__(self, db_file: Path) -> None:
+        self._db_file = db_file
+        self.create_table()
+
+    def create_table(self) -> None:
+        """Create table."""
+        if self._db_file.exists():
+            return
+
+        try:  # noqa: WPS229
+            connection = sqlite3.connect(self._db_file)
+            sqlite_create_table_query = """CREATE TABLE session_table (
+                                        primary_key TEXT NOT NULL,
+                                        sort_key INTEGER NOT NULL,
+                                        data_key TEXT NOT NULL,
+                                        PRIMARY KEY (primary_key, sort_key)
+                                        );"""
+
+            cursor = connection.cursor()
+            cursor.execute(sqlite_create_table_query)
+            connection.commit()
+            logger.info("SQLite table created")
+
+            cursor.close()
+
+        except sqlite3.Error as error:
+            logger.exception("Error while creating a sqlite table", error)
+        finally:
+            if connection:
+                connection.close()
 
     def healthcheck(self) -> bool:
         """Verify that the DB can be accessed and that it is ready."""
-        dynamodb_client = boto3.client(
-            "dynamodb", region_name=self._resource_region  # pyright: ignore
-        )
-
         try:
-            dynamodb_client.describe_table(TableName=self._table_name)
-        except dynamodb_client.exceptions.ResourceNotFoundException:
-            logger.exception("Cannot find DynamoDB table")
+            sqlite3.connect(self._db_file)
+        except Exception:
+            logger.exception("Cannot find db table")
             return False
 
         return True
 
     def add_session_turn(self, session_turn: SimBotSessionTurn) -> None:
         """Add a session turn to the table."""
-        try:
-            response = self._table.put_item(
-                Item={
-                    self.primary_key: session_turn.session_id,
-                    self.sort_key: session_turn.idx,
-                    self.data_key: session_turn.json(by_alias=True),
-                },
-                ConditionExpression="attribute_not_exists(#sort_key)",
-                ExpressionAttributeNames={"#sort_key": self.sort_key},
-            )
-            logger.debug(response)
-        except ClientError as err:
-            logger.exception("Could not add turn to table.")
+        try:  # noqa: WPS229
+            connection = sqlite3.connect(self._db_file)
+            cursor = connection.cursor()
 
-            error_code = err.response["Error"]["Code"]  # pyright: ignore
-            if error_code != "ConditionalCheckFailedException":
-                raise err
+            sqlite_insert_with_param = """INSERT OR REPLACE INTO session_table
+                                (primary_key, sort_key, data_key)
+                                VALUES (?, ?, ?);"""
+
+            data_tuple = (
+                session_turn.session_id,
+                session_turn.idx,
+                session_turn.json(by_alias=True),
+            )
+            cursor.execute(sqlite_insert_with_param, data_tuple)
+            connection.commit()
+            logger.info("Successfully inserted turn into table")
+
+            cursor.close()
+
+        except sqlite3.Error as error:
+            logger.exception("Failed to insert turn into table")
+            raise error
+        finally:
+            if connection:
+                connection.close()
 
     def put_session_turn(self, session_turn: SimBotSessionTurn) -> None:
         """Put a session turn to the table.
 
         If the turn already exists, it WILL overwrite it.
         """
-        try:
-            self._table.put_item(
-                Item={
-                    self.primary_key: session_turn.session_id,
-                    self.sort_key: session_turn.idx,
-                    self.data_key: session_turn.json(by_alias=True),
-                },
-            )
-        except ClientError as err:
-            logger.exception("Could not add turn to table.")
-            raise err
+        self.add_session_turn(session_turn)
 
     def get_session_turn(self, session_id: str, idx: int) -> SimBotSessionTurn:
         """Get the session turn from the table."""
-        try:
-            response = self._table.get_item(Key={self.primary_key: session_id, self.sort_key: idx})
-        except ClientError as err:
-            logger.exception("Could not get session turn from table")
-            raise err
+        try:  # noqa: WPS229
+            connection = sqlite3.connect(self._db_file)
+            cursor = connection.cursor()
 
-        return SimBotSessionTurn.parse_obj(response["Item"][self.data_key])
+            sql_select_query = "select * from session_table where primary_key = ? and sort_key = ?"
+            cursor.execute(sql_select_query, (session_id, idx))
+            turn = cursor.fetchone()
+            cursor.close()
+
+        except sqlite3.Error as error:
+            logger.exception("Failed to read data from table")
+            raise error
+        finally:
+            if connection:
+                connection.close()
+
+        return SimBotSessionTurn.parse_raw(turn[2])
 
     def get_all_session_turns(self, session_id: str) -> list[SimBotSessionTurn]:
         """Get all the turns for a given session."""
         try:
             all_raw_turns = self._get_all_session_turns(session_id)
-        except ClientError as query_err:
+        except Exception as query_err:
             logger.exception("Could not query for session turns")
             raise query_err
 
@@ -92,7 +124,7 @@ class SimBotSessionDbClient(DynamoDbClient):
                 parsed_responses = list(
                     thread_pool.map(
                         SimBotSessionTurn.parse_raw,
-                        (response_item[self.data_key] for response_item in all_raw_turns),
+                        (response_item[2] for response_item in all_raw_turns),
                     )
                 )
             except Exception:
@@ -108,18 +140,23 @@ class SimBotSessionDbClient(DynamoDbClient):
 
         return sorted_responses
 
-    def _get_all_session_turns(self, session_id: str) -> list[dict[str, Any]]:
-        response = self._table.query(KeyConditionExpression=Key(self.primary_key).eq(session_id))
+    def _get_all_session_turns(self, session_id: str) -> list[tuple[str, int, str]]:
+        try:  # noqa: WPS229
+            connection = sqlite3.connect(self._db_file)
+            cursor = connection.cursor()
 
-        all_response_items = response["Items"]
-
-        # If not all the instances have been returned, get the next set
-        while "LastEvaluatedKey" in response:
-            response = self._table.query(
-                KeyConditionExpression=Key(self.primary_key).eq(session_id),
-                ExclusiveStartKey=response["LastEvaluatedKey"],
+            sql_select_query = (
+                "select * from session_table where primary_key = ? ORDER BY sort_key"
             )
+            cursor.execute(sql_select_query, (session_id,))
+            turns = cursor.fetchall()
+            cursor.close()
 
-            all_response_items.extend(response["Items"])
+        except sqlite3.Error as error:
+            logger.exception("Failed to read data from table")
+            raise error
+        finally:
+            if connection:
+                connection.close()
 
-        return all_response_items  # type: ignore[unreachable]
+        return turns
